@@ -30,78 +30,78 @@
     "storybook": "start-storybook -p 6006"
   },
   "devDependencies": {
-    "@babel/cli": "^7.15.7",
-    "@babel/core": "^7.15.8",
-    "@babel/eslint-parser": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.15.0",
-    "@storybook/addon-actions": "^6.3.12",
-    "@storybook/addon-essentials": "^6.3.12",
-    "@storybook/addon-links": "^6.3.12",
-    "@storybook/builder-webpack5": "^6.3.12",
-    "@storybook/react": "^6.3.12",
-    "@testing-library/jest-dom": "^5.14.1",
+    "@babel/cli": "^7.17.0",
+    "@babel/core": "^7.17.0",
+    "@babel/eslint-parser": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/preset-typescript": "^7.16.7",
+    "@storybook/addon-actions": "^6.4.18",
+    "@storybook/addon-essentials": "^6.4.18",
+    "@storybook/addon-links": "^6.4.18",
+    "@storybook/builder-webpack5": "^6.4.18",
+    "@storybook/react": "^6.4.18",
+    "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "@types/react-helmet": "^6.1.4",
-    "@typescript-eslint/eslint-plugin": "^5.2.0",
-    "@typescript-eslint/parser": "^5.2.0",
+    "@types/react-helmet": "^6.1.5",
+    "@typescript-eslint/eslint-plugin": "^5.11.0",
+    "@typescript-eslint/parser": "^5.11.0",
     "axios-mock-adapter": "^1.20.0",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.3",
     "babel-plugin-module-resolver": "^4.1.0",
-    "css-loader": "^6.4.0",
-    "dotenv-webpack": "^7.0.3",
-    "eslint": "^8.1.0",
+    "css-loader": "^6.6.0",
+    "dotenv-webpack": "^7.1.0",
+    "eslint": "^8.8.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-jest": "^25.2.2",
+    "eslint-config-react-app": "^7.0.0",
+    "eslint-plugin-jest": "^26.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^5.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-testing-library": "^5.0.5",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
     "husky": "^7.0.4",
     "i": "^0.3.7",
-    "jest": "^27.3.1",
-    "mini-css-extract-plugin": "^2.4.3",
-    "node-sass": "^6.0.1",
-    "prettier": "^2.4.1",
+    "jest": "^27.5.0",
+    "mini-css-extract-plugin": "^2.5.3",
+    "node-sass": "^7.0.1",
+    "prettier": "^2.5.1",
     "react-test-renderer": "^17.0.2",
-    "resolve-url-loader": "^4.0.0",
-    "sass-loader": "^12.2.0",
-    "shx": "^0.3.3",
+    "resolve-url-loader": "^5.0.0",
+    "sass-loader": "^12.4.0",
+    "shx": "^0.3.4",
     "style-loader": "^3.3.1",
     "thread-loader": "^3.0.4",
     "ts-loader": "^9.2.6",
-    "typescript": "^4.4.4",
-    "webpack": "^5.60.0",
-    "webpack-assets-manifest": "^5.0.6",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.3.1"
+    "typescript": "^4.5.5",
+    "webpack": "^5.68.0",
+    "webpack-assets-manifest": "^5.1.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4"
   },
   "dependencies": {
-    "@popperjs/core": "^2.10.2",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.33",
-    "@types/react-bootstrap": "^0.32.28",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-router": "^5.1.17",
-    "@types/react-router-dom": "^5.3.2",
-    "axios": "^0.24.0",
+    "@popperjs/core": "^2.11.2",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.15",
+    "@types/react": "^17.0.39",
+    "@types/react-bootstrap": "^0.32.29",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router": "^5.1.18",
+    "@types/react-router-dom": "^5.3.3",
+    "axios": "^0.25.0",
     "bootstrap": "^5.1.3",
     "draft-js": "^0.11.7",
-    "express": "^4.17.1",
-    "prop-types": "^15.7.2",
+    "express": "^4.17.2",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0",
+    "react-bootstrap": "^2.1.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.3.0"
+    "react-router-dom": "^6.2.1"
   },
   "browserslist": {
     "production": [
